commander/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.5.7
azure-core==1.25.1
azure-eventhub==5.10.1
azure-eventhub-checkpointstoreblob==1.1.4
azure-iot-device==2.12.0
azure-iot-hub==2.6.0
backcall==0.2.0
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
deprecation==2.1.0
dj-database-url==0.5.0
Django==4.0
entrypoints==0.4
executing==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
install==1.3.5
ipykernel==6.16.0
ipython==8.5.0
isodate==0.6.1
isort==5.10.1
janus==1.0.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
msrest==0.7.1
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.5
oauthlib==3.2.1
packaging==21.3
paho-mqtt==1.6.1
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.6
uamqp==1.6.0
Unipath==1.1
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==5.3.0
zope.interface==5.5.2
